server/requirements.txt
websockets>=12.0,<13.0
asyncio>=3.4.3
pyyaml>=6.0
wyoming>=1.4.0
protobuf>=4.21.0
zeroconf>=0.131.0
aiohttp
